+++ src/cli/build.rs
@@ -51,7 +51,11 @@ pub struct Build {
 	#[arg(short, long)]
 	ts: bool,
 
-	/// Spawn the Argon child process
+	/// Run Argon asynchroniously
+	#[arg(short = 'A', long = "async")]
+	run_async: bool,
+
+	/// Spawn the Argon child process (internal)
 	#[arg(long, hide = true)]
 	argon_spawn: bool,
 }
@@ -60,7 +64,7 @@ impl Build {
 	pub fn main(self) -> Result<()> {
 		let config = Config::new();
 
-		if self.watch && !self.argon_spawn && config.run_async {
+		if self.watch && !self.argon_spawn && (self.run_async || config.run_async) {
 			return self.spawn();
 		}