+++ src/cli/serve.rs
@@ -43,7 +43,11 @@ pub struct Serve {
 	#[arg(short, long)]
 	ts: bool,
 
-	/// Spawn the Argon child process
+	/// Run Argon asynchroniously
+	#[arg(short = 'A', long = "async")]
+	run_async: bool,
+
+	/// Spawn the Argon child process (internal)
 	#[arg(long, hide = true)]
 	argon_spawn: bool,
 }
@@ -52,7 +56,7 @@ impl Serve {
 	pub fn main(self) -> Result<()> {
 		let config = Config::new();
 
-		if !self.argon_spawn && config.run_async {
+		if !self.argon_spawn && (self.run_async || config.run_async) {
 			return self.spawn();
 		}